requirements.txt
Django>=6.0,<7.0
Pillow>=12.0
django-environ>=0.13
gunicorn>=25.0
whitenoise>=6.0
